solved merge conflict with sgct and ghoul

+++ modules/globebrowsing/src/tileprovider.cpp
@@ -868,7 +868,7 @@ TemporalTileProvider::TemporalTileProvider(const ghoul::Dictionary& dictionary)
         useFixedTime = dictionary.value<bool>(temporal::UseFixedTimeInfo.identifier);
     }
     addProperty(useFixedTime);
-    
+
     if (dictionary.hasValue<std::string>(temporal::FixedTimeInfo.identifier)) {
         fixedTime = dictionary.value<std::string>(temporal::FixedTimeInfo.identifier);
     }
@@ -891,7 +891,9 @@ bool initialize(TileProvider& tp) {
 
     ghoul_assert(!tp.isInitialized, "TileProvider can only be initialized once.");
 
-    if (TileProvider::NumTileProviders > std::numeric_limits<uint16_t>::max() - 1) {
+    if (TileProvider::NumTileProviders >
+        static_cast<unsigned int>(std::numeric_limits<uint16_t>::max()) - 1)
+    {
         LERRORC(
             "TileProvider",
             "Number of tile providers exceeds 65535. Something will break soon"
